+++ TASMA/Page/GradePage.xaml.cs
@@ -14,8 +14,9 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using TASMA.Database;
 using TASMA.DataInterfaces;
+using TASMA.MessageBox;
 
-namespace TASMA
+namespace TASMA.Pages
 {
     /// <summary>
     /// 학년 페이지 입니다.
@@ -126,9 +127,26 @@ namespace TASMA
         /// <param name="e">NULL</param>
         private void OnDeleteData(object sender, EventArgs e)
         {
+            if (DataRectangleManager.IsModified == false)
+                return;
+
             var dataRect = sender as DataRectangle;
-            adminDAO.DeleteGrade(dataRect.Data);
-            Invalidate();           
+
+            var passwordConfirm = new TasmaPasswordMessageBox("Input password", "Input password to delete grade");
+            passwordConfirm.ShowDialog();
+
+            if (passwordConfirm.IsDetermined) {
+
+                if (adminDAO.Authenticate(passwordConfirm.Input))
+                {
+                    adminDAO.DeleteGrade(dataRect.Data);
+                    Invalidate();
+                }else
+                {
+                    var alert = new TasmaAlertMessageBox("Incorrect password", "Password doesn't match");
+                    alert.ShowDialog();
+                }
+            }            
         }
 
         /// <summary>
@@ -153,7 +171,8 @@ namespace TASMA
                     Invalidate();
                 }else
                 {
-                    MessageBox.Show("Grade already exists");
+                    var alert = new TasmaAlertMessageBox("Duplication", "Grade already exists");
+                    alert.ShowDialog();
                     return;
                 }
             }
